feat(helm): support for empty global.imageRegistry in common.images.image
Refactor the `common.images.image` Helm template to handle the scenario when `global.imageRegistry` is not set. The template checks if `.imageRoot.registry` is not an empty string before using it as a custom registry name. If `.imageRoot.registry` is not set or empty, the template defaults to `.global.imageRegistry`. Additionally, the template now returns the image name without the registry section if `$registryName` is empty. These changes improve the reliability and flexibility of the Helm chart.

BREAKING CHANGE: This change may affect any deployment that relies on the previous behavior of `common.images.image`. If `global.imageRegistry` is not set and `.imageRoot.registry` is empty, the image name generated by the template will default to not including the registry section.